+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/weekly/verticaltotal/worktime/KrcdtWekAggrPremTime.java
@@ -40,7 +40,7 @@ public class KrcdtWekAggrPremTime extends ContractUkJpaEntity implements Seriali
 	public int premiumTime;
 	/** 割増金額 */
 	@Column(name = "PREMIUM_AMOUNT")
-	public int premiumAmount;
+	public long premiumAmount;
 
 	/** マッチング：週別実績の勤怠時間 */
 	@ManyToOne
